Add Fulfillment microservice that have two modules order and dispatch module with some endpoints for progress

+++ src/operations/role/role.useCase.impl.ts
@@ -4,6 +4,7 @@ import { RoleCreateDto, RoleUpdateDto } from './role.entity';
 import { RoleUseCases } from './role.useCase';
 import { IPagination } from '../../common/types';
 import { RoleRepository } from './role.repository';
+import { RpcException } from '@nestjs/microservices';
 
 @Injectable()
 export class RoleUseCaseImpl implements RoleUseCases {
@@ -26,18 +27,54 @@ export class RoleUseCaseImpl implements RoleUseCases {
     pageSize: number,
     search?: string,
   ): Promise<{ roles: Partial<Role>[]; pagination: IPagination }> {
-    return this.roleRepo.findAllRoles(page, pageSize, search);
+    const skip = (page - 1) * pageSize;
+    const where: any = {};
+
+    if (search) {
+      where.OR = [{ name: { contains: search, mode: 'insensitive' } }];
+    }
+
+    const [roles, total] = await this.roleRepo.findAllRoles(
+      skip,
+      pageSize,
+      where,
+    );
+
+    const totalPages = Math.ceil(total / pageSize);
+
+    return {
+      roles,
+      pagination: {
+        total,
+        page,
+        pageSize,
+        totalPages,
+      },
+    };
   }
 
-  async deleteRole(payload: { id?: string; name?: string }): Promise<string> {
-    if (payload.id) {
-      return this.roleRepo.deleteById(payload.id);
-    } else if (payload.name) {
-      return this.roleRepo.deleteByName(payload.name);
-    } else {
-      return 'Invalid payload: Must provide either id or name';
-    }
+ async deleteRole(payload: { id?: string; name?: string }): Promise<string> {
+  console.log("payload: ", payload);
+
+  if (payload.id) {
+    const role = await this.roleRepo.findRolById(payload.id);
+    if (!role) throw new RpcException(`Role ${payload.id} not found`);
+
+    await this.roleRepo.deleteById(payload.id);
+    return `Role deleted successfully with id: ${payload.id}`;
   }
+
+  if (payload.name) {
+    const role = await this.roleRepo.findRoleByName(payload.name);
+    if (!role) throw new RpcException(`Role ${payload.name} not found`);
+
+    await this.roleRepo.deleteByName(payload.name);
+    return `Role deleted successfully with name: ${payload.name}`;
+  }
+
+  throw new RpcException('Invalid payload: Must provide either id or name');
+}
+
 
   async updateRole(id: string, data: RoleUpdateDto): Promise<Role> {
     return this.roleRepo.updateRole(id, data);
